+++ CHANGELOG.rst
@@ -37,6 +37,20 @@ the author interprets the terms below as follows:
   API is stable.
   The package is available on PyPI (and perhaps conda).
 
+x.y.z - 2020-xx-xx (dev)
+------------------------
+Changed
+^^^^^^^
+* add_lorentzians moved from nmr.plt to nmr.math.
+  Besides being more appropriate here,
+  it fixes a problem where importing from nmrsim.plt also requires importing
+  from matplotlib/Tkinter.
+  Some environments
+  (some Unix systems; BeeWare's Briefcase packaging tool)
+  do not have Tkinter available.
+* nmrsim.plt will substitute the matplotlib "Agg" backend for "TkAgg"
+  if tkinter is not found on the users system, and print a warning message
+  that plots will not be visible.
 
 0.4.0 - 2020-xx-xx (beta release)
 ---------------------------------